Title: 정수 두 개의 크기 비교, Time: 289ms, Memory: 17MB, Status: Wrong Answer - CodeTree

+++ 241003/정수 두 개의 크기 비교/compare-the-size-of-two-integers.java
@@ -22,8 +22,8 @@ public class Main {
         System.out.println(answer);
     }
 
-    static int bs(int i) {
-        int answer = i;
+    static long bs(int i) {
+        long answer = i;
         int left = 0;
         int right = i-1;
         while (left <= right) {
